Refactoring UIController
